Extract NaCl-specific entry point in C++ Example
Split the pp_module.cc source file into two parts: NaCl-specific
"entry_point_nacl.cc" and toolchain-independent "application.{h|cc}".

This is a preparatory CL for implementing a different entry point
for Emscripten/WebAssembly builds (as tracked by #220).
